ASC.Web.Core: StudioWhatsNewNotify: refactoring 48b941f368c768811c4859b20d25441618dc3acc
# Conflicts:
#	products/ASC.Files/Core/Configuration/ProductEntryPoint.cs
#	products/ASC.Files/Server/ProductEntryPoint.cs
#	web/ASC.Web.Core/IProduct.cs
#	web/ASC.Web.Core/Notify/StudioWhatsNewNotify.cs
#	web/ASC.Web.Core/Product.cs
