+++ web/ASC.Web.Core/Product.cs
@@ -49,7 +49,7 @@ public abstract class Product : IProduct
 
     public virtual void Shutdown() { }
 
-    public virtual Task<IEnumerable<ActivityInfo>> GetAuditEventsAsync(DateTime scheduleDate, Guid userId, Tenant tenant, WhatsNewType whatsNewType, ILogger logger)
+    public virtual Task<IEnumerable<ActivityInfo>> GetAuditEventsAsync(DateTime scheduleDate, Guid userId, Tenant tenant, WhatsNewType whatsNewType)
     {
         return Task.FromResult(Enumerable.Empty<ActivityInfo>());
     }
